backend/requirements.txt
# ===== Backend Python — Tech Infographic Generator =====

# --- API Server ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# --- Data models ---
pydantic>=2.6.0

# --- Image rendering (PIL) ---
Pillow>=10.2.0
cairosvg>=2.7.0

# --- LLM text analysis ---
anthropic>=0.43.0
openai>=1.14.0
httpx>=0.27.0

# --- Config ---
python-dotenv>=1.0.0

# --- CLI (generate.py) ---
click>=8.1.0
rich>=13.7.0
